Refactoring and Comments
