+++ tests/tests_ui/test_impedance_frame.py
@@ -1,28 +1,15 @@
-import pytest
-import PySide6
-from time import sleep
-
-from PySide6 import QtWidgets, QtCore
-from pytestqt.qtbot import QtBot
-from PySide6.QtTest import QTest, QSignalSpy
-from PySide6.QtCore import Qt, QTimer
 from PySide6.QtWidgets import QApplication, QMessageBox
-import unittest
-from unittest.mock import Mock, MagicMock, patch
+from pytestqt.qtbot import QtBot
+from PySide6.QtCore import Qt, QTimer
+from unittest.mock import Mock
 from exploredesktop.modules.app_settings import ImpModes
 from exploredesktop.modules.explore_interface import ExploreInterface
 from collections import namedtuple
 from explorepy.settings_manager import SettingsManager
-from test_Window import connect_device
+from test_window import connect_device
 from exploredesktop.modules.app_settings import Messages
-from unittest.mock import patch
 
 
-# TODO tests todo
-# TODO check if filters have been applied
-# TODO test upper hotbar for evrything output file and stuff
-# TODO last week for Andrea
-# Wednesday vorbeikommen MEntalab anschrieben
 
 def get_device_mock():
     e_interface = ExploreInterface()
@@ -149,57 +136,53 @@ def navigate_to_impedance_view(qtbot, window):
     qtbot.mouseClick(imp_button, Qt.LeftButton)
 
 
-class TestImpedance:
-    # Can only always test one test as each test needs its own device...
+def test_impedance_modes(qtbot):
+    window = connect_device(qtbot)
+    # window.show()
+    imp = window.imp_frame
+    navigate_to_impedance_view(qtbot, window)
+    # press dropdown and select Wet Electrodes
+    drop_down = imp.ui.imp_mode
+    qtbot.addWidget(drop_down)
+    qtbot.keyClicks(drop_down, ImpModes.WET.value)
+    # Press measure btn
+    meas_btn = imp.ui.btn_imp_meas
+    qtbot.mouseClick(meas_btn, Qt.LeftButton)
+    qtbot.wait(1000)
+    # Check the color
+    # get Impframe view call get_stylesheet check color?
+    imp_model = imp.get_graph().model
+    assert imp_model.mode == ImpModes.WET
+    # disable Impedance measurement
+    qtbot.mouseClick(meas_btn, Qt.LeftButton)
+    # Change to Dry Electrodes
+    qtbot.keyClicks(drop_down, ImpModes.DRY.value)
+    assert imp_model.mode == ImpModes.DRY
+    # stop measuring Impedance
+    qtbot.mouseClick(meas_btn, Qt.LeftButton)
 
-    """
-    def test_impedance_modes(self, qtbot):
-        window = connect_device(qtbot)
-        # window.show()
-        imp = window.imp_frame
-        navigate_to_impedance_view(qtbot, window)
-        # press dropdown and select Wet Electrodes
-        drop_down = imp.ui.imp_mode
-        qtbot.addWidget(drop_down)
-        qtbot.keyClicks(drop_down, ImpModes.WET.value)
-        # Press measure btn
-        meas_btn = imp.ui.btn_imp_meas
-        qtbot.mouseClick(meas_btn, Qt.LeftButton)
-        qtbot.wait(1000)
-        # Check the color
-        # get Impframe view call get_stylesheet check color?
-        imp_model = imp.get_graph().model
-        assert imp_model.mode == ImpModes.WET
-        # disable Impedance measurement
-        qtbot.mouseClick(meas_btn, Qt.LeftButton)
-        # Change to Dry Electrodes
-        qtbot.keyClicks(drop_down, ImpModes.DRY.value)
-        assert imp_model.mode == ImpModes.DRY
-        # stop measuring Impedance
-        qtbot.mouseClick(meas_btn, Qt.LeftButton)
-    """
 
-    def test_pop_up(self, qtbot):
-        # Helper method where Qtbot connects the device and returns QMainWindow instance
-        window = connect_device(qtbot)
-        # Only for visualisation
-        window.show()
-        # Helper method where Qtbot navigates to the Impedance View Frame
-        navigate_to_impedance_view(qtbot, window)
-        imp = window.imp_frame
-        # Get the reference to the question mark button in the Impedance Frame
-        info = imp.ui.imp_meas_info
-        # Give the reference to qtbot so it know it
-        qtbot.addWidget(info) 
-        
-        def handle_dialog():
-            # Get an instance of the currently open window
-            messagebox = QtWidgets.QApplication.activeWindow()
-            assert messagebox.text() == Messages.IMP_INFO
-            ok_button = messagebox.button(QtWidgets.QMessageBox.Ok)
-            qtbot.mouseClick(ok_button, Qt.LeftButton)
+def test_info_pop_up(qtbot):
+    # Helper method where Qtbot connects the device and returns QMainWindow instance
+    window = connect_device(qtbot)
+    # Only for visualisation
+    window.show()
+    # Helper method where Qtbot navigates to the Impedance View Frame
+    navigate_to_impedance_view(qtbot, window)
+    imp = window.imp_frame
+    # Get the reference to the question mark button in the Impedance Frame
+    info = imp.ui.imp_meas_info
+    # Give the reference to qtbot so it know it
+    qtbot.addWidget(info)
 
-        QtCore.QTimer.singleShot(100, handle_dialog)
-        qtbot.mouseClick(info, Qt.LeftButton, delay=1)
+    def handle_dialog():
+        # Get an instance of the currently open window and answer it
+        messagebox = QApplication.activeWindow()
+        assert messagebox.text() == Messages.IMP_INFO
+        ok_button = messagebox.button(QMessageBox.Ok)
+        qtbot.mouseClick(ok_button, Qt.LeftButton)
+
+    QTimer.singleShot(100, handle_dialog)
+    qtbot.mouseClick(info, Qt.LeftButton, delay=1)
