+++ tests/tests_ui/test_window.py
@@ -1,0 +1,28 @@
+from pytestqt.qtbot import QtBot
+from PySide6.QtTest import QTest
+import exploredesktop.main_window as mw
+from PySide6.QtCore import Qt
+from time import sleep
+
+
+# Move this to test_BlueToothFrame?
+def connect_device(qtbot):
+    window = mw.MainWindow()
+    bt = window.bt_frame
+    input_field = bt.ui.dev_name_input
+    qtbot.addWidget(input_field)
+    qtbot.keyClicks(input_field, "855E")
+    qtbot.keyClick(input_field, Qt.Key_Enter)
+    sleep(5)
+    qtbot.wait(5000)
+    return window
+
+
+def test_close_window(qtbot):
+    window = mw.MainWindow()
+    qtbot.addWidget(window)
+    window.show()
+    assert window.isVisible()
+    QTest.qWaitForWindowExposed(window)
+    window.close()
+    assert not window.isVisible()
